st-llama/requirements.txt
streamlit
langchain
llama-cpp-python
chromadb